☕POST/GET/DELETE skills:   - skillController   - Integer -> Long во всех репозиториях   - skillDto

+++ src/main/java/com/example/StudyFriends/exceptions/ResourceNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.StudyFriends.exceptions;
 
 public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String resource, Integer id) {
+    public ResourceNotFoundException(String resource, Long id) {
         super(resource + " с ID " + id + " не найден");
     }
